+++ container-engine-lib/lib/backend_impls/docker/docker_kurtosis_backend/logs_database_functions/implementations/loki/loki_container_config_provider.go
@@ -29,7 +29,7 @@ func newLokiContainerConfigProvider(config *LokiConfig, httpPortNumber uint16) *
 }
 
 func (loki *lokiContainerConfigProvider) GetPrivateHttpPortSpec() (*port_spec.PortSpec, error) {
-	privateHttpPortSpec, err := port_spec.NewPortSpec(loki.httpPortNumber, httpPortProtocol)
+	privateHttpPortSpec, err := port_spec.NewPortSpec(loki.httpPortNumber, httpPortProtocol, "")
 	if err != nil {
 		return nil, stacktrace.Propagate(
 			err,